Heart Diseases Detector
The "Heart Disease Prediction App" is a user-friendly application built using Tkinter and TensorFlow. It leverages a neural network model trained on a heart disease dataset to predict the likelihood of heart disease based on user-inputted medical data. The app features an organized and visually appealing interface, allowing users to input relevant information and receive predictions with ease. The inclusion of a copyright label adds a professional touch to the application. This project combines machine learning, graphical user interface development, and data visualization to create a practical tool for predicting heart disease risk.